requirements.txt
discord-py
discord-py-interactions
requests
pynacl
ffmpeg
bs4
discord-py-slash-command